debug

+++ train.py
@@ -31,7 +31,7 @@ class Trainer:
         self.start_time = time.time()
 
         # Initialize learning rate schedule
-        epoch_milestones = np.array([self.opt.nEpochs * i for i in self.opt.schedule]) 
+        epoch_milestones = np.array([int(self.opt.nEpochs * i) for i in self.opt.schedule]) 
         self.scheduler = torch.optim.lr_scheduler.MultiStepLR(self.optimizer, epoch_milestones, gamma=0.1) 
 
     def train(self, epoch):
